+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
 - `Instance.create_instance` method now takes list of tuples (index, internal float value) when generating sparse instances
 - added `weka.core.database` module for loading data from a database
 - added `make_copy` class method to `Clusterer` class
+- added `make_copy` class method to `Associator` class
 - ...